Data.ByteString.Optics

Documentation

class IsByteString t where

Traversals for ByteStrings.

Methods

packedBytes :: Iso' [Word8] t

pack (or unpack) a list of bytes into a strict or lazy ByteString.

pack x ≡ x ^. packedBytes
unpack x ≡ x ^. re packedBytes
packedBytes ≡ re unpackedBytes

packedChars :: Iso' String t

pack (or unpack) a list of characters into a strict or lazy ByteString.

When writing back to the ByteString it is assumed that every Char lies between 'x00' and 'xff'.

pack x ≡ x ^. packedChars
unpack x ≡ x ^. re packedChars
packedChars ≡ re unpackedChars

bytes :: IxTraversal' Int64 t Word8

Traverse each Word8 in a strict or lazy ByteString

This Traversal walks each strict ByteString chunk in a tree-like fashion enable zippers to seek to locations more quickly and accelerate many monoidal queries, but up to associativity (and constant factors) it is equivalent to the much slower:

bytes ≡ unpackedBytes . traversed

anyOf bytes (== 0x80) :: ByteString -> Bool

chars :: IxTraversal' Int64 t Char

Traverse the individual bytes in a strict or lazy ByteString as characters.

When writing back to the ByteString it is assumed that every Char lies between 'x00' and 'xff'.

This Traversal walks each strict ByteString chunk in a tree-like fashion enable zippers to seek to locations more quickly and accelerate many monoidal queries, but up to associativity (and constant factors) it is equivalent to the much slower:

chars ≡ unpackedChars . traversed

anyOf chars (== 'c') :: ByteString -> Bool

unpackedBytes :: IsByteString t => Iso' t [Word8]

unpack (or pack) a ByteString into a list of bytes.

unpackedBytes ≡ re packedBytes
unpack x ≡ x ^. unpackedBytes
pack x ≡  x ^. re unpackedBytes

unpackedBytes :: Iso' ByteString [Word8]
unpackedBytes :: Iso' ByteString [Word8]

unpackedChars :: IsByteString t => Iso' t String

unpack (or pack) a list of characters into a strict (or lazy) ByteString

When writing back to the ByteString it is assumed that every Char lies between 'x00' and 'xff'.

unpackedChars ≡ re packedChars
unpack x ≡ x ^. unpackedChars
pack x ≡ x ^. re unpackedChars

unpackedChars :: Iso' ByteString String
unpackedChars :: Iso' ByteString String

pattern Bytes :: IsByteString t => [Word8] -> t

pattern Chars :: IsByteString t => [Char] -> t