Data.Text.Strict.Optics

Description

This module provides Isos for converting strict Text to or from a String or Builder, and an IxTraversal for traversing the individual characters of a Text.

If you need to work with both strict and lazy text, Data.Text.Optics provides combinators that support both varieties using a typeclass.

Documentation

packed :: Iso' String Text

This isomorphism can be used to pack (or unpack) strict Text.

>>> "hello" ^. packed -- :: Text
"hello"

pack x ≡ x ^. packed
unpack x ≡ x ^. re packed
packed ≡ re unpacked
packed ≡ iso pack unpack

unpacked :: Iso' Text String

This isomorphism can be used to unpack (or pack) strict Text.

>>> Strict.pack "hello" ^. unpacked -- :: String
"hello"

This Iso is provided for notational convenience rather than out of great need, since

unpacked ≡ re packed

pack x ≡ x ^. re unpacked
unpack x ≡ x ^. packed
unpacked ≡ iso unpack pack

builder :: Iso' Text Builder

Convert between strict Text and Builder .

fromText x ≡ x ^. builder
toStrict (toLazyText x) ≡ x ^. re builder

text :: IxTraversal' Int Text Char

Traverse the individual characters in strict Text.

>>> anyOf text (=='o') (Strict.pack "hello")
True

When the type is unambiguous, you can also use the more general each.

text ≡ unpacked % traversed
text ≡ each

Note that when just using this as a Setter, sets map can be more efficient.

utf8 :: Prism' ByteString Text

Encode/Decode a strict Text to/from strict ByteString, via UTF-8.

>>> utf8 # Strict.pack "☃"
"\226\152\131"

_Text :: Iso' Text String

This is an alias for unpacked that makes it more obvious how to use it with #

> _Text # "hello" -- :: Text

"hello"

pattern Text :: String -> Text