Optics.At

Description

This module provides optics for Map and Set-like containers, including an AffineTraversal to traverse a key in a map or an element of a sequence:

>>> preview (ix 1) ['a','b','c']
Just 'b'

a Lens to get, set or delete a key in a map:

>>> set (at 0) (Just 'b') (Map.fromList [(0, 'a')])
fromList [(0,'b')]

and a Lens to insert or remove an element of a set:

>>> IntSet.fromList [1,2,3,4] & contains 3 .~ False
fromList [1,2,4]

This module includes the core definitions from Optics.At.Core along with extra (orphan) instances.

Type families

type family Index s

Type family that takes a key-value container type and returns the type of keys (indices) into the container, for example Index (Map k a) ~ k. This is shared by Ixed, At and Contains.

type family IxValue m

Type family that takes a key-value container type and returns the type of values stored in the container, for example IxValue (Map k a) ~ a. This is shared by both Ixed and At.

Ixed

class Ixed m where

Provides a simple AffineTraversal lets you traverse the value at a given key in a Map or element at an ordinal position in a list or Seq.

Minimal complete definition

Nothing

Associated Types

type IxKind m

Type family that takes a key-value container type and returns the kind of optic to index into it. For most containers, it's An_AffineTraversal, Representable (Naperian) containers it is A_Lens, and multi-maps would have A_Traversal.

type IxKind m = An_AffineTraversal

Methods

ix :: Index m -> Optic' (IxKind m) NoIx m (IxValue m)

NB: Setting the value of this AffineTraversal will only set the value in at if it is already present.

If you want to be able to insert missing values, you want at.

>>> [1,2,3,4] & ix 2 %~ (*10)
[1,2,30,4]

>>> "abcd" & ix 2 .~ 'e'
"abed"

>>> "abcd" ^? ix 2
Just 'c'

>>> [] ^? ix 2
Nothing

ixAt :: At m => Index m -> AffineTraversal' m (IxValue m)

A definition of ix for types with an At instance. This is the default if you don't specify a definition for ix.

At

class (Ixed m, IxKind m ~ An_AffineTraversal) => At m where

At provides a Lens that can be used to read, write or delete the value associated with a key in a Map-like container on an ad hoc basis.

An instance of At should satisfy:

ix k ≡ at k % _Just

Methods

at :: Index m -> Lens' m (Maybe (IxValue m))

>>> Map.fromList [(1,"world")] ^. at 1
Just "world"

>>> at 1 ?~ "hello" $ Map.empty
fromList [(1,"hello")]

Note: Usage of this function might introduce space leaks if you're not careful to make sure that values put inside the Just constructor are evaluated. To force the values and avoid such leaks, use at' instead.

Note: Map-like containers form a reasonable instance, but not Array-like ones, where you cannot satisfy the Lens laws.

at' :: At m => Index m -> Lens' m (Maybe (IxValue m))

Version of at strict in the value inside the Just constructor.

Example:

>>> (at () .~ Just (error "oops") $ Nothing) `seq` ()
()

>>> (at' () .~ Just (error "oops") $ Nothing) `seq` ()
*** Exception: oops
...

>>> view (at ()) (Just $ error "oops") `seq` ()
()

>>> view (at' ()) (Just $ error "oops") `seq` ()
*** Exception: oops
...

It also works as expected for other data structures:

>>> (at 1 .~ Just (error "oops") $ Map.empty) `seq` ()
()

>>> (at' 1 .~ Just (error "oops") $ Map.empty) `seq` ()
*** Exception: oops
...

sans :: At m => Index m -> m -> m

Delete the value associated with a key in a Map-like container

sans k = at k .~ Nothing

Contains

class Contains m where

This class provides a simple Lens that lets you view (and modify) information about whether or not a container contains a given Index. Instances are provided for Set-like containers only.

Methods

contains :: Index m -> Lens' m Bool

>>> IntSet.fromList [1,2,3,4] ^. contains 3
True

>>> IntSet.fromList [1,2,3,4] ^. contains 5
False

>>> IntSet.fromList [1,2,3,4] & contains 3 .~ False
fromList [1,2,4]

Orphan instances

Ixed ByteString

Instance details

Associated Types

type IxKind ByteString

Methods

ix :: Index ByteString -> Optic' (IxKind ByteString) NoIx ByteString (IxValue ByteString)

Ixed ByteString

Instance details

Associated Types

type IxKind ByteString

Methods

ix :: Index ByteString -> Optic' (IxKind ByteString) NoIx ByteString (IxValue ByteString)

Ixed Text

Instance details

Associated Types

type IxKind Text

Methods

ix :: Index Text -> Optic' (IxKind Text) NoIx Text (IxValue Text)

Ixed Text

Instance details

Associated Types

type IxKind Text

Methods

ix :: Index Text -> Optic' (IxKind Text) NoIx Text (IxValue Text)

(Eq k, Hashable k) => At (HashSet k)

Instance details

Methods

at :: Index (HashSet k) -> Lens' (HashSet k) (Maybe (IxValue (HashSet k)))

(Eq a, Hashable a) => Contains (HashSet a)

Instance details

Methods

contains :: Index (HashSet a) -> Lens' (HashSet a) Bool

(Eq k, Hashable k) => Ixed (HashSet k)

Instance details

Associated Types

type IxKind (HashSet k)

Methods

ix :: Index (HashSet k) -> Optic' (IxKind (HashSet k)) NoIx (HashSet k) (IxValue (HashSet k))

Ixed (Vector a)

Instance details

Associated Types

type IxKind (Vector a)

Methods

ix :: Index (Vector a) -> Optic' (IxKind (Vector a)) NoIx (Vector a) (IxValue (Vector a))

Prim a => Ixed (Vector a)

Instance details

Associated Types

type IxKind (Vector a)

Methods

ix :: Index (Vector a) -> Optic' (IxKind (Vector a)) NoIx (Vector a) (IxValue (Vector a))

Storable a => Ixed (Vector a)

Instance details

Associated Types

type IxKind (Vector a)

Methods

ix :: Index (Vector a) -> Optic' (IxKind (Vector a)) NoIx (Vector a) (IxValue (Vector a))

Unbox a => Ixed (Vector a)

Instance details

Associated Types

type IxKind (Vector a)

Methods

ix :: Index (Vector a) -> Optic' (IxKind (Vector a)) NoIx (Vector a) (IxValue (Vector a))

(Eq k, Hashable k) => At (HashMap k a)

Instance details

Methods

at :: Index (HashMap k a) -> Lens' (HashMap k a) (Maybe (IxValue (HashMap k a)))

(Eq k, Hashable k) => Ixed (HashMap k a)

Instance details

Associated Types

type IxKind (HashMap k a)

Methods

ix :: Index (HashMap k a) -> Optic' (IxKind (HashMap k a)) NoIx (HashMap k a) (IxValue (HashMap k a))