Optics.Empty

Description

This module defines the AsEmpty class, which provides a Prism for a type that may be _Empty.

>>> isn't _Empty [1,2,3]
True

>>> case Nothing of { Empty -> True; _ -> False }
True

Documentation

class AsEmpty a where

Class for types that may be _Empty.

Minimal complete definition

Nothing

Methods

_Empty :: Prism' a ()

>>> isn't _Empty [1,2,3]
True

pattern Empty :: AsEmpty a => a

Pattern synonym for matching on any type with an AsEmpty instance.

>>> case Nothing of { Empty -> True; _ -> False }
True

Orphan instances

AsEmpty ByteString

Instance details

Methods

_Empty :: Prism' ByteString ()

AsEmpty ByteString

Instance details

Methods

_Empty :: Prism' ByteString ()

AsEmpty Text

Instance details

Methods

_Empty :: Prism' Text ()

AsEmpty Text

Instance details

Methods

_Empty :: Prism' Text ()

AsEmpty (HashSet a)

Instance details

Methods

_Empty :: Prism' (HashSet a) ()

AsEmpty (Vector a)

Instance details

Methods

_Empty :: Prism' (Vector a) ()

Storable a => AsEmpty (Vector a)

Instance details

Methods

_Empty :: Prism' (Vector a) ()

Unbox a => AsEmpty (Vector a)

Instance details

Methods

_Empty :: Prism' (Vector a) ()

AsEmpty (HashMap k a)

Instance details

Methods

_Empty :: Prism' (HashMap k a) ()