Optics.Extra.Internal.ByteString

Description

This module spends a lot of time fiddling around with ByteString internals to work around http://hackage.haskell.org/trac/ghc/ticket/7556 on older Haskell Platforms and to improve constant and asymptotic factors in our performance.

Documentation

traversedStrictTree :: IxTraversal' Int64 ByteString Word8

Traverse a strict ByteString in a relatively balanced fashion, as a balanced tree with biased runs of elements at the leaves.

traversedStrictTree8 :: IxTraversal' Int64 ByteString Char

Traverse a strict ByteString in a relatively balanced fashion, as a balanced tree with biased runs of elements at the leaves, pretending the bytes are chars.

traversedLazy :: IxTraversal' Int64 ByteString Word8

An IxTraversal of the individual bytes in a lazy ByteString.

traversedLazy8 :: IxTraversal' Int64 ByteString Char

An IxTraversal of the individual bytes in a lazy ByteString pretending the bytes are chars.