Optics.Extra.Internal.Zoom

Zoom

newtype Focusing m c s

Used by Zoom to zoom into StateT.

Constructors

Focusing

Fields

unfocusing :: m (c, s)

stateZoom :: (Is k A_Lens, Monad m) => Optic' k is t s -> (s -> m (c, s)) -> t -> m (c, t)

stateZoomMaybe :: (Is k An_AffineTraversal, Monad m) => Optic' k is t s -> (s -> m (c, s)) -> t -> m (Maybe c, t)

stateZoomMany :: (Is k A_Traversal, Monad m, Monoid c) => Optic' k is t s -> (s -> m (c, s)) -> t -> m (c, t)

newtype FocusingWith w m c s

Used by Zoom to zoom into RWST.

Constructors

FocusingWith

Fields

unfocusingWith :: m (c, s, w)

rwsZoom :: (Is k A_Lens, Monad m) => Optic' k is t s -> (r -> s -> m (c, s, w)) -> r -> t -> m (c, t, w)

rwsZoomMaybe :: (Is k An_AffineTraversal, Monad m, Monoid w) => Optic' k is t s -> (r -> s -> m (c, s, w)) -> r -> t -> m (Maybe c, t, w)

rwsZoomMany :: (Is k A_Traversal, Monad m, Monoid w, Monoid c) => Optic' k is t s -> (r -> s -> m (c, s, w)) -> r -> t -> m (c, t, w)

newtype May a

Make a Monoid out of Maybe for error handling.

Constructors

May

Fields

getMay :: Maybe a

shuffleMay :: Maybe (May c) -> May (Maybe c)

newtype Err e a

Make a Monoid out of Either for error handling.

Constructors

Err

Fields

getErr :: Either e a

shuffleErr :: Maybe (Err e c) -> Err e (Maybe c)

Magnify

newtype Effect m r

Wrap a monadic effect.

Constructors

Effect

Fields

getEffect :: m r

newtype EffectRWS w s m c

Wrap a monadic effect. Used when magnifying RWST.

Constructors

EffectRWS

Fields

getEffectRWS :: s -> m (c, s, w)

rwsMagnify :: Is k A_Getter => Optic' k is a b -> (b -> s -> f (c, s, w)) -> a -> s -> f (c, s, w)

rwsMagnifyMaybe :: (Is k An_AffineFold, Applicative m, Monoid w) => Optic' k is a b -> (b -> s -> m (c, s, w)) -> a -> s -> m (Maybe c, s, w)

rwsMagnifyMany :: (Is k A_Fold, Monad m, Monoid w, Monoid c) => Optic' k is a b -> (b -> s -> m (c, s, w)) -> a -> s -> m (c, s, w)

Misc

shuffleS :: s -> Maybe (c, s) -> (Maybe c, s)

shuffleW :: Monoid w => Maybe (c, w) -> (Maybe c, w)