Optics.View

Description

EXPERIMENTAL

Documentation

class ViewableOptic k r where

Generalized view (even more powerful than view from the lens library).

View the value(s) pointed to by an optic.

The type of the result depends on the optic. You get:

Exactly one result a with Iso, Lens, ReversedPrism and Getter.

At most one result Maybe a with Prism, AffineTraversal and AffineFold.

Monoidal summary of all results Monoid a => a with Traversal and Fold.

When in doubt, use specific, flavour restricted versions. This function is mostly useful for things such as passthrough.

Associated Types

type ViewResult k r :: Type

Methods

gview :: MonadReader s m => Optic' k is s r -> m (ViewResult k r)

gviews :: MonadReader s m => Optic' k is s a -> (a -> r) -> m (ViewResult k r)

guse :: (ViewableOptic k a, MonadState s m) => Optic' k is s a -> m (ViewResult k a)

Use the target of a Lens, Iso, or Getter in the current state, or use a summary of a Fold or Traversal that points to a monoidal value.

>>> evalState (guse _1) ('a','b')
'a'

>>> evalState (guse _2) ("hello","world")
"world"

Since: 0.2

guses :: (ViewableOptic k r, MonadState s m) => Optic' k is s a -> (a -> r) -> m (ViewResult k r)

Use the target of a Lens, Iso or Getter in the current state, or use a summary of a Fold or Traversal that points to a monoidal value.

>>> evalState (guses _1 length) ("hello","world")
5

Since: 0.2

glistening :: (ViewableOptic k r, MonadWriter s m) => Optic' k is s r -> m a -> m (a, ViewResult k r)

This is a generalized form of listen that only extracts the portion of the log that is focused on by a Getter. If given a Fold or a Traversal then a monoidal summary of the parts of the log that are visited will be returned.

Since: 0.2

glistenings :: (ViewableOptic k r, MonadWriter s m) => Optic' k is s a -> (a -> r) -> m b -> m (b, ViewResult k r)

This is a generalized form of listen that only extracts the portion of the log that is focused on by a Getter. If given a Fold or a Traversal then a monoidal summary of the parts of the log that are visited will be returned.

Since: 0.2