Optics.Zoom

Zoom

class (MonadState s m, MonadState t n) => Zoom m n s t | m -> s, n -> t, m t -> n, n s -> m where

This class allows us to zoom in, changing the State supplied by many different monad transformers, potentially quite deep in a monad transformer stack.

Its functions can be used to run a monadic action in a larger State than it was defined in, using a Lens', an AffineTraversal' or a Traversal'.

This is commonly used to lift actions in a simpler State Monad into a State Monad with a larger State type.

When used with a Traversal' over multiple values, the actions for each target are executed sequentially and the results are aggregated.

This can be used to edit pretty much any Monad transformer stack with a State in it!

>>> flip L.evalState ('a','b') $ zoom _1 $ use equality
'a'

>>> flip S.execState ('a','b') $ zoom _1 $ equality .= 'c'
('c','b')

>>> flip L.execState [(1,2),(3,4)] $ zoomMany traversed $ _2 %= (*10)
[(1,20),(3,40)]

>>> flip S.runState [('a',"b"),('c',"d")] $ zoomMany traversed $ _2 <%= (\x -> x <> x)
("bbdd",[('a',"bb"),('c',"dd")])

>>> flip S.evalState ("a","b") $ zoomMany each (use equality)
"ab"

Methods

zoom :: Is k A_Lens => Optic' k is t s -> m c -> n c infixr 2

zoomMaybe :: Is k An_AffineTraversal => Optic' k is t s -> m c -> n (Maybe c) infixr 2

zoomMany :: (Is k A_Traversal, Monoid c) => Optic' k is t s -> m c -> n c infixr 2

Magnify

class (MonadReader b m, MonadReader a n) => Magnify m n b a | m -> b, n -> a, m a -> n, n b -> m where

This class allows us to magnify part of the environment, changing the environment supplied by many different Monad transformers. Unlike zoom this can change the environment of a deeply nested Monad transformer.

Its functions can be used to run a monadic action in a larger environment than it was defined in, using a Getter or an AffineFold.

They act like local, but can in many cases change the type of the environment as well.

They're commonly used to lift actions in a simpler Reader Monad into a Monad with a larger environment type.

They can be used to edit pretty much any Monad transformer stack with an environment in it:

>>> (1,2) & magnify _2 (+1)
3

>>> flip runReader (1,2) $ magnify _1 ask
1

>>> flip runReader (1,2,[10..20]) $ magnifyMaybe (_3 % _tail) ask
Just [11,12,13,14,15,16,17,18,19,20]

Methods

magnify :: Is k A_Getter => Optic' k is a b -> m c -> n c infixr 2

magnifyMaybe :: Is k An_AffineFold => Optic' k is a b -> m c -> n (Maybe c) infixr 2

class (MonadReader b m, MonadReader a n, Magnify m n b a) => MagnifyMany m n b a | m -> b, n -> a, m a -> n, n b -> m where

Extends Magnify with an ability to magnify using a Fold over multiple targets so that actions for each one are executed sequentially and the results are aggregated.

There is however no sensible instance of MagnifyMany for StateT.

Methods

magnifyMany :: (Is k A_Fold, Monoid c) => Optic' k is a b -> m c -> n c infixr 2